+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/verifier
 
+## 0.11.16
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.11.16
+
 ## 0.11.15
 
 ### Patch Changes